# Improve performance for const field