This improves the performance for const fields while having the performance for mutable fields nearly equal to before.

Incredibly enough this beats base mutable structs by a huge amount!!

```julia
julia> using ProtoStructs

julia> @proto mutable struct A
           const x::Int
           const y::Int
       end

julia> mutable struct B
           const x::Int
           const y::Int
       end

julia> v_a = [A(x, x) for x in 1:1000];

julia> v_b = [B(x, x) for x in 1:1000];

julia> using BenchmarkTools

julia> @benchmark sum(x.y for x in $v_a)
BenchmarkTools.Trial: 10000 samples with 700 evaluations.
 Range (min … max):  177.974 ns … 397.629 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     183.216 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   186.304 ns ±  11.263 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

    ▃▆██▇▆▄▂▁ ▁▃▃▃▂▂▂▁▁                                         ▂
  ▂█████████████████████████▇▇▆▆▅▆▆▇▆▄▄▆▆▄▅▄▆▅▅▄▅▆▅▆▆▅▆▆▇▆▇▆▅▂▅ █
  178 ns        Histogram: log(frequency) by time        233 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> @benchmark sum(x.y for x in $v_b)
BenchmarkTools.Trial: 10000 samples with 57 evaluations.
 Range (min … max):  871.544 ns …   4.771 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     872.877 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   889.405 ns ± 114.206 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █                                                             ▁
  █▄▄▃▃▁▁▄▄▃▁▄▄▆▇▇▆▅▅▆▆▆▆▇▄▅▅▅▅▅▄▅▃▅▅▅▅▅▅▄▃▄▄▄▃▄▄▁▁▅▁▃▄▄▄▃▃▄▄▃▄ █
  872 ns        Histogram: log(frequency) by time       1.44 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

Almost looks like something to say to some Julia core developers, but I assume there are probably good reasons why the perf difference is that big