@Tortar 2023-10-18
Actually it seems like it works that well only when **all** fields are constants, but still maybe it is good to have also this opt