@Tortar 2023-10-18
Actually (2) with more real benchmarks I see almost no improvement, and also I still see the proto mutable version lagging behind the normal version by a lot (10x)